requirements.txt
pandas
numpy
scikit-learn
streamlit
pillow
easyocr
joblib
plotly
